requirements.txt
requests==2.32.4
pyspark==3.4.4
python-dotenv==1.0.1
snowflake-connector-python==3.14.0
